Moved solver classes into a solver package
