+++ src/main/java/solvers/SolverThreadWrapper.java
